put back label utils not required by trymerge anymore
