+++ .github/scripts/test_check_labels.py
@@ -1,16 +1,17 @@
 """test_check_labels.py"""
 
-from typing import Any, TYPE_CHECKING
+from typing import Any, List
 from unittest import TestCase, mock, main
 
-from check_labels import main as check_labels_main
-from label_utils import LABEL_ERR_MSG
-from test_trymerge import mock_gh_get_info
-
-# TODO: this is a temp workaround to avoid circular dependencies,
-#       and should be removed once GitHubPR is refactored out of trymerge script.
-if TYPE_CHECKING:
-    from trymerge import GitHubPR
+from check_labels import (
+    main as check_labels_main,
+    add_label_err_comment,
+    delete_all_label_err_comments,
+)
+from github_utils import GitHubComment
+from label_utils import BOT_AUTHORS, LABEL_ERR_MSG, LABEL_ERR_MSG_TITLE
+from test_trymerge import mocked_gh_graphql, mock_gh_get_info
+from trymerge import GitHubPR
 
 def mock_parse_args() -> object:
     class Object(object):
@@ -24,8 +25,63 @@ def mock_add_label_err_comment(pr: "GitHubPR") -> None:
 def mock_delete_all_label_err_comments(pr: "GitHubPR") -> None:
     pass
 
+def mock_get_comments() -> List[GitHubComment]:
+    return [
+        # Case 1 - a non label err comment
+        GitHubComment(
+            body_text="mock_body_text",
+            created_at="",
+            author_login="",
+            author_association="",
+            editor_login=None,
+            database_id=1,
+        ),
+        # Case 2 - a label err comment
+        GitHubComment(
+            body_text=" #" + LABEL_ERR_MSG_TITLE,
+            created_at="",
+            author_login=BOT_AUTHORS[1],
+            author_association="",
+            editor_login=None,
+            database_id=2,
+        ),
+    ]
+
 
 class TestCheckLabels(TestCase):
+    @mock.patch('trymerge.gh_graphql', side_effect=mocked_gh_graphql)
+    @mock.patch('trymerge.GitHubPR.get_comments', return_value=[mock_get_comments()[0]])
+    @mock.patch('check_labels.gh_post_pr_comment')
+    def test_correctly_add_label_err_comment(
+        self, mock_gh_post_pr_comment: Any, mock_get_comments: Any, mock_gh_grphql: Any
+    ) -> None:
+        "Test add label err comment when similar comments don't exist."
+        pr = GitHubPR("pytorch", "pytorch", 75095)
+        add_label_err_comment(pr)
+        mock_gh_post_pr_comment.assert_called_once()
+
+    @mock.patch('trymerge.gh_graphql', side_effect=mocked_gh_graphql)
+    @mock.patch('trymerge.GitHubPR.get_comments', return_value=[mock_get_comments()[1]])
+    @mock.patch('check_labels.gh_post_pr_comment')
+    def test_not_add_label_err_comment(
+        self, mock_gh_post_pr_comment: Any, mock_get_comments: Any, mock_gh_grphql: Any
+    ) -> None:
+        "Test not add label err comment when similar comments exist."
+        pr = GitHubPR("pytorch", "pytorch", 75095)
+        add_label_err_comment(pr)
+        mock_gh_post_pr_comment.assert_not_called()
+
+    @mock.patch('trymerge.gh_graphql', side_effect=mocked_gh_graphql)
+    @mock.patch('trymerge.GitHubPR.get_comments', return_value=mock_get_comments())
+    @mock.patch('check_labels.gh_post_delete_comment')
+    def test_correctly_delete_all_label_err_comments(
+        self, mock_gh_post_delete_comment: Any, mock_get_comments: Any, mock_gh_grphql: Any
+    ) -> None:
+        "Test only delete label err comment."
+        pr = GitHubPR("pytorch", "pytorch", 75095)
+        delete_all_label_err_comments(pr)
+        mock_gh_post_delete_comment.assert_called_once_with("pytorch", "pytorch", 2)
+
     @mock.patch('trymerge.gh_get_pr_info', return_value=mock_gh_get_info())
     @mock.patch('check_labels.parse_args', return_value=mock_parse_args())
     @mock.patch('check_labels.has_required_labels', return_value=False)
