+++ .github/scripts/test_label_utils.py
@@ -1,12 +1,7 @@
-from typing import Any, List
+from typing import Any
 from unittest import TestCase, mock, main
 
-from github_utils import GitHubComment
 from label_utils import (
-    BOT_AUTHORS,
-    LABEL_ERR_MSG_TITLE,
-    add_label_err_comment,
-    delete_all_label_err_comments,
     get_last_page_num_from_header,
     gh_get_labels,
     has_required_labels,
@@ -18,28 +13,6 @@ from test_trymerge import mocked_gh_graphql
 release_notes_labels = [
     "release notes: nn",
 ]
-
-def mock_get_comments() -> List[GitHubComment]:
-    return [
-        # Case 1 - a non label err comment
-        GitHubComment(
-            body_text="mock_body_text",
-            created_at="",
-            author_login="",
-            author_association="",
-            editor_login=None,
-            database_id=1,
-        ),
-        # Case 2 - a label err comment
-        GitHubComment(
-            body_text=" #" + LABEL_ERR_MSG_TITLE,
-            created_at="",
-            author_login=BOT_AUTHORS[1],
-            author_association="",
-            editor_login=None,
-            database_id=2,
-        ),
-    ]
 
 class TestLabelUtils(TestCase):
     MOCK_HEADER_LINKS_TO_PAGE_NUMS = {
@@ -96,39 +69,6 @@ class TestLabelUtils(TestCase):
         "Test PR with 'topic: not user facing' label"
         pr = GitHubPR("pytorch", "pytorch", 75095)
         self.assertTrue(has_required_labels(pr))
-
-    @mock.patch('trymerge.gh_graphql', side_effect=mocked_gh_graphql)
-    @mock.patch('trymerge.GitHubPR.get_comments', return_value=[mock_get_comments()[0]])
-    @mock.patch('label_utils.gh_post_pr_comment')
-    def test_correctly_add_label_err_comment(
-        self, mock_gh_post_pr_comment: Any, mock_get_comments: Any, mock_gh_grphql: Any
-    ) -> None:
-        "Test add label err comment when similar comments don't exist."
-        pr = GitHubPR("pytorch", "pytorch", 75095)
-        add_label_err_comment(pr)
-        mock_gh_post_pr_comment.assert_called_once()
-
-    @mock.patch('trymerge.gh_graphql', side_effect=mocked_gh_graphql)
-    @mock.patch('trymerge.GitHubPR.get_comments', return_value=[mock_get_comments()[1]])
-    @mock.patch('label_utils.gh_post_pr_comment')
-    def test_not_add_label_err_comment(
-        self, mock_gh_post_pr_comment: Any, mock_get_comments: Any, mock_gh_grphql: Any
-    ) -> None:
-        "Test not add label err comment when similar comments exist."
-        pr = GitHubPR("pytorch", "pytorch", 75095)
-        add_label_err_comment(pr)
-        mock_gh_post_pr_comment.assert_not_called()
-
-    @mock.patch('trymerge.gh_graphql', side_effect=mocked_gh_graphql)
-    @mock.patch('trymerge.GitHubPR.get_comments', return_value=mock_get_comments())
-    @mock.patch('label_utils.gh_post_delete_comment')
-    def test_correctly_delete_all_label_err_comments(
-        self, mock_gh_post_delete_comment: Any, mock_get_comments: Any, mock_gh_grphql: Any
-    ) -> None:
-        "Test only delete label err comment."
-        pr = GitHubPR("pytorch", "pytorch", 75095)
-        delete_all_label_err_comments(pr)
-        mock_gh_post_delete_comment.assert_called_once_with("pytorch", "pytorch", 2)
 
 
 if __name__ == "__main__":
